Brojne izmjene

+++ Admin/Modules/Projects.aspx.cs
@@ -9,24 +9,33 @@ using System.Web;
 using System.Web.Services;
 using System.Web.UI;
 using System.Web.UI.WebControls;
+using System.Data.Entity;
+using System.Globalization;
+using MLE.Admin.program;
 
 namespace MLE.Admin.Modules
 {
     public partial class Projects : System.Web.UI.Page
     {
-        private int projectId = 0;
+        protected List<DB.Type> Types = new List<DB.Type>();
+        protected List<DB.User> Users = new List<DB.User>();
+        protected int projectId = 0;
+        public int Pages = 0;
+        private int Skip = 0;
+        public int current_page = 1;
+        private bool check_page = true;
 
         protected void Page_Load(object sender, EventArgs e)
         {
-            PopulateRepeater();
+            PopulateRepeater(true);
         }
 
         protected void Page_PreRender(object sender, EventArgs e)
         {
             if (Request.QueryString["id"] != null)
             {
-                var str = Request.QueryString["id"].FirstOrDefault().ToString();
-                projectId = int.Parse(str);
+                var str = Request.QueryString["id"].ToString();
+                int.TryParse(str, out projectId);
 
                 if (projectId == 0)
                     btnDelete.Visible = false;
@@ -34,11 +43,45 @@ namespace MLE.Admin.Modules
             }
         }
 
-        private List<Project> GetAllProjects()
+        private void GetTypesOnProject(MLEEntities db, Project project)
+        {
+            var e = project.Example.ToList();
+            var AllTypes = new List<int>();
+            foreach (var _ in e)
+                AllTypes.Add(_.TypeId ?? 0);
+
+            AllTypes.Distinct();
+
+            Types = db.Type.Where(x => AllTypes.Contains(x.Id)).ToList();
+        }
+
+        private void GetUsersOnProject(MLEEntities db, Project project)
+        {
+            var u = project.UserProject.Select(x => x.UserId).ToList();
+            Users = db.User.Where(x => u.Contains(x.Id)).ToList();
+        }
+
+        private List<Project> GetAllProjects(bool update_pager)
         {
             using (var db = new MLEEntities())
             {
-                return db.Project.ToList();
+                var projects = db.Project.Include(x => x.Example).ToList();
+                var count = projects.Count;
+                var default_by_page = 20;
+                current_page = 1;
+                if (check_page)
+                    if (Request.QueryString["page"] != null)
+                        int.TryParse(Request.QueryString["page"].ToString(), out current_page);
+
+                Pages = (int)Math.Ceiling((double)count / (double)default_by_page);
+                current_page = current_page > Pages ? Pages : current_page;
+                Skip = (current_page - 1) * default_by_page;
+                if (update_pager)
+                    Pager.CreatePager(Pages, current_page, phPager, "Projects");
+
+                projects = projects.Skip(Skip).Take(default_by_page).ToList();
+
+                return projects;
             }
         }
 
@@ -46,27 +89,47 @@ namespace MLE.Admin.Modules
         {
             using (var db = new MLEEntities())
             {
-                Project _dbProject =  db.Project.Where(p => p.Id == id).FirstOrDefault();
+                Project _dbProject = db.Project.Where(p => p.Id == id).FirstOrDefault();
 
-                if(_dbProject != null)
+                PopulateDropDowns(db);
+
+                if (_dbProject != null)
                 {
+                    GetTypesOnProject(db, _dbProject);
+                    GetUsersOnProject(db, _dbProject);
                     txtName.Text = _dbProject.Name != null ? _dbProject.Name : "";
                     txtDescription.Text = _dbProject.Description != null ? _dbProject.Description : "";
                     txtSpentTime.Text = _dbProject.TimeSpent != null ? _dbProject.TimeSpent.ToString() : "";
                     txtDateCreated.Text = _dbProject.DateCreated != null ? _dbProject.DateCreated.Value.ToString("dd.MM.yyyy") : "";
                     txtStartDate.Text = _dbProject.Start_Date != null ? _dbProject.Start_Date.Value.ToString("dd.MM.yyyy") : "";
                     txtEndDate.Text = _dbProject.End_Date != null ? _dbProject.End_Date.Value.ToString("dd.MM.yyyy") : "";
-                    txtStatus.Text = _dbProject.StatusId != null ? _dbProject.StatusId.ToString() : "";
+                    ddlStatus.SelectedValue = _dbProject.StatusId != null ? _dbProject.StatusId.Value.ToString() : "1";
                 }
             }
         }
 
-        private List<Project> PopulateRepeater()
+        private void PopulateDropDowns(MLEEntities db)
         {
-            rpt.DataSource = GetAllProjects();
-            rpt.DataBind();
+            ddlStatus.DataSource = db.Status.ToList();
+            ddlStatus.DataValueField = "Id";
+            ddlStatus.DataTextField = "Name";
+            ddlStatus.DataBind();
 
-            return GetAllProjects();
+            ddlType.DataSource = db.Type.Where(x => x.Active == true).ToList();
+            ddlType.DataValueField = "Id";
+            ddlType.DataTextField = "Name";
+            ddlType.DataBind();
+
+            ddlUser.DataSource = db.User.Where(x => x.IsActive == true).ToList();
+            ddlUser.DataValueField = "Id";
+            ddlUser.DataTextField = "Username";
+            ddlUser.DataBind();
+        }
+
+        private void PopulateRepeater(bool update_pager)
+        {
+            rpt.DataSource = GetAllProjects(update_pager);
+            rpt.DataBind();
         }
 
         private void Save()
@@ -76,7 +139,6 @@ namespace MLE.Admin.Modules
                 DateTime.TryParse(txtStartDate.Text, out DateTime start_date);
 
                 DateTime.TryParse(txtEndDate.Text, out DateTime end_date);
-                int.TryParse(txtStatus.Text, out int status);
 
                 Project _project = new Project()
                 {
@@ -86,7 +148,7 @@ namespace MLE.Admin.Modules
                     TimeSpent = TimeSpan.Zero,
                     Start_Date = start_date,
                     End_Date = end_date.Year == 0001 ? DateTime.Now : end_date,
-                    StatusId = status
+                    StatusId = int.Parse(ddlStatus.SelectedValue)
                 };
 
                 db.Project.Attach(_project);
@@ -108,11 +170,11 @@ namespace MLE.Admin.Modules
 
                     _dbProject.Name = txtName.Text;
                     _dbProject.Description = txtDescription.Text;
-                    _dbProject.DateCreated = DateTime.Parse(txtDateCreated.Text);
+                    //_dbProject.DateCreated = DateTime.ParseExact(txtDateCreated.Text, "dd/MM/yyyy", null);
                     _dbProject.TimeSpent = TimeSpan.Zero;
-                    _dbProject.Start_Date = DateTime.Parse(txtStartDate.Text);
-                    _dbProject.End_Date = DateTime.Parse(txtStartDate.Text);
-                    _dbProject.StatusId = 2;
+                    //_dbProject.Start_Date = DateTime.ParseExact(txtStartDate.Text, "dd/MM/yyyy", null);
+                    //_dbProject.End_Date = DateTime.ParseExact(txtStartDate.Text, "dd/MM/yyyy", null);
+                    _dbProject.StatusId = int.Parse(ddlStatus.SelectedValue);
 
                     db.SaveChanges();
                 }
@@ -121,26 +183,77 @@ namespace MLE.Admin.Modules
 
         private void Delete()
         {
-            using (var db = new MLEEntities())
+            if (Request.QueryString["id"] != null)
             {
-                Project _dbProject = db.Project.Where(u => u.Id == projectId).FirstOrDefault();
-                db.Project.Attach(_dbProject);
-                db.Project.Remove(_dbProject);
+                var id = Request.QueryString["id"].ToString();
+                projectId = int.Parse(id);
+                using (var db = new MLEEntities())
+                {
+                    var examples = db.Example.Where(x => x.ProjectId == projectId).ToList();
+                    foreach (var item in examples)
+                    {
+                        var marked = db.Marked.Where(x => x.ExampleId == item.Id).ToList();
+                        foreach (var i in marked)
+                        {
+                            db.Marked.Attach(i);
+                            db.Marked.Remove(i);
+                            db.SaveChanges();
+                        }
 
-                db.SaveChanges();
+                        var ec = db.ExampleCategory.Where(x => x.ExampleId == item.Id).ToList();
+                        foreach (var i in ec)
+                        {
+                            db.ExampleCategory.Attach(i);
+                            db.ExampleCategory.Remove(i);
+                            db.SaveChanges();
+                        }
+
+                        db.Example.Attach(item);
+                        db.Example.Remove(item);
+                        db.SaveChanges();
+                    }
+
+                    var userProject = db.UserProject.Where(x => x.ProjectId == projectId).ToList();
+                    foreach (var item in userProject)
+                    {
+                        db.UserProject.Attach(item);
+                        db.UserProject.Remove(item);
+                        db.SaveChanges();
+                    }
+
+                    Project _dbProject = db.Project.Where(u => u.Id == projectId).FirstOrDefault();
+                    db.Project.Attach(_dbProject);
+                    db.Project.Remove(_dbProject);
+                    db.SaveChanges();
+                }
+
+                Response.Redirect("Projects.aspx");
             }
         }
 
         protected void btnSave_Click(object sender, EventArgs e)
         {
-            Save();
+            if (Request.QueryString["id"] != null)
+            {
+                var id = Request.QueryString["id"].ToString();
+                if (id != "")
+                {
+                    projectId = int.Parse(id);
+                    if (projectId == 0)
+                        Save();
+                    else
+                        Update();
+                }
+            }
+            PopulateRepeater(false);
         }
 
         protected void btnDelete_Click(object sender, EventArgs e)
         {
             Delete();
+            Response.Redirect("Projects.aspx");
         }
-        
+
         private void ExportToJson(int projectId)
         {
             IList<Example> _projectExampleList = null;
@@ -157,7 +270,7 @@ namespace MLE.Admin.Modules
                 {
                     _dbMarkedData = db.Marked.Where(m => m.ExampleId == item.Id).ToList();
 
-                    for(int i = 0; i < _dbMarkedData.Count; i++)
+                    for (int i = 0; i < _dbMarkedData.Count; i++)
                     {
                         int subcategoryId = _dbMarkedData[i].SubcategoryId.HasValue ? _dbMarkedData[i].SubcategoryId.Value : 1;
 
@@ -193,7 +306,7 @@ namespace MLE.Admin.Modules
             string fileName = dbProject.Name + " - json - " + DateTime.Now.ToString("yyyy-MM-dd") + ".txt";
             string filePath = Path.Combine(HttpContext.Current.Server.MapPath("~/JsonData/"), fileName);
 
-            File.WriteAllText(filePath, json); 
+            File.WriteAllText(filePath, json);
 
             FileStream fs = null;
 
@@ -202,20 +315,85 @@ namespace MLE.Admin.Modules
             byte[] byteFile = new byte[fs.Length];
             fs.Read(byteFile, 0, Convert.ToInt32(fs.Length));
             fs.Close();
-            
+
             HttpContext context = HttpContext.Current;
             context.Response.Buffer = true;
             context.Response.Clear();
             context.Response.AddHeader("content-disposition", "attachment; filename=" + fileName);
             context.Response.BinaryWrite(byteFile);
-            context.Response.End();  
+            context.Response.End();
         }
 
         protected void btnExport_Click(object sender, EventArgs e)
         {
             int projectID = int.Parse(Request.QueryString["id"].ToString());
-            
-            ExportToJson(projectID);          
+
+            ExportToJson(projectID);
+        }
+
+        protected void btnAddType_Click(object sender, EventArgs e)
+        {
+            projectId = int.Parse(Request.QueryString["id"].ToString());
+            var typeId = int.Parse(ddlType.SelectedValue);
+
+            using (var db = new MLEEntities())
+            {
+                var t = db.Example.Where(x => x.ProjectId == projectId).ToList();
+                foreach (var item in t)
+                {
+                    db.Example.Attach(item);
+                    item.TypeId = typeId;
+                    db.SaveChanges();
+                }
+            }
+        }
+
+        protected void btnRemoveType_Click(object sender, EventArgs e)
+        {
+            projectId = int.Parse(Request.QueryString["id"].ToString());
+            var typeId = int.Parse(ddlType.SelectedValue);
+
+            using (var db = new MLEEntities())
+            {
+                var t = db.Example.Where(x => x.ProjectId == projectId).ToList();
+                foreach (var item in t)
+                {
+                    db.Example.Attach(item);
+                    item.TypeId = null;
+                    db.SaveChanges();
+                }
+            }
+        }
+
+        protected void btnRemoveUser_Click(object sender, EventArgs e)
+        {
+            projectId = int.Parse(Request.QueryString["id"].ToString());
+            var userId = int.Parse(ddlUser.SelectedValue);
+
+            using (var db = new MLEEntities())
+            {
+                var up = db.UserProject.FirstOrDefault(x => x.ProjectId == projectId && x.UserId == userId);
+                if(up != null)
+                {
+                    db.UserProject.Remove(up);
+                    db.SaveChanges();
+                }
+            }
+        }
+
+        protected void btnAddUser_Click(object sender, EventArgs e)
+        {
+            projectId = int.Parse(Request.QueryString["id"].ToString());
+            var userId = int.Parse(ddlUser.SelectedValue);
+
+            using (var db = new MLEEntities())
+            {
+                var up = new UserProject();
+                up.ProjectId = projectId;
+                up.UserId = userId;
+                db.UserProject.Add(up);
+                db.SaveChanges();
+            }
         }
     }
 }
